Fluid iframes, https-fixes and added PlexRequests
Added function to measure height of viewport and subtract the height of
the navigation tabs and then resize the iframes based on that (so that
the iframes always cover the entire height of the viewport without
causing scrollbars).

Also removed "http://" from Google Fonts URL, which breaks https
(unsecured resources). Replaced with simply "//", which works in both
cases.

Also added Plex Requests link.