Started a new fork with a bunch of updates.
Renamed to MTPHP (Managethis PHP) to signify the difference and minimize
confusion.

Opted to do this outside of Managethis which is already good for novice
users that might not want to fiddle with PHP.

- Added logic for the generation of menu items and iframes.
- Moved configuration to a separate config-file that will not be
overwritten if you update from this repository.
- Added ability to enable/disable items.

To come:
- Dynamically generated landing-pages without code-reuse.
- Password protection of MTPHP

Will add more functionality as I find the time and as long as it's fun.