Implemented dynamically generated landing pages
Reworked the code for the landing page as well, as the form
submit/window.location method causes a questionmark (?) to be appended
to the URL, which results in a 404 on nzbget and possibly other
services.

Removed the need to specify "fa-lg" and "fa-3x" in icon classes.

Removed static landing-pages, moved CSS to css directory and changed
path to background to images/