Add Tab Colors
Closes mescon/Muximux#92

Initial implementation.  Want to look at settings spacing with the new
element, possible re-layout.  But, that delves into how responsive the
cards are in mobile now.   Also, not crazy about current color picker
implementation, spectrum has a better one, but it wasn't cooperating, so
that's what's there now.

But...it works.