Add Mobile Override Button
Fixes mescon/Muximux#68

Add a settings item for a "Mobile override button".  When enabled, and
when Muximux is rendering in Mobile mode, the button will be displayed
next to the refresh button.

When pressed, the button toggles between rendering for desktop and
mobile mode, causing app tiles to be displayed in the top bar.