Add methods necessary for git-less updating
Add methods to download a list of branches and their SHA from github,
along with a method to download a zip of the current branch and extract
it.

TODO:  Need methods to save list to settings so we're not always calling
it, parsing out the current SHA on install/settings create/no SHA value
in settings.  Need to generate the list of branches from the hidden
setting for branches.

Finally, add a new mechanism that runs on page load to detect what the
SHA is of the current branch vs. the one online, and if it's different,
notify that an update is available.

Also, need a handler to ask if they would like to switch codebase to
whatever branch they pick in settings.