Whoa
Yeah.  Um.  Where to start...

More work on settings menu layout and responsiveness.  Wound up recoding
most of parse_ini so it was easier to move bits about.  Everything seems
to be good except for the branch listing function, which I'll look at
tomorrow.

Also, it seems to like to scrap the branches from Github more than it
should, and I think it might also be downloading updates for the branch
at random.

Added dynamic color bar!  On mobile, the bar color will switch to
correspond to the selected tab.  COOL.

Added overflow mechanism to bar so that having more tabs than the screen
is wide will result in the extra tabs being dumped in the dropdown.  I
have not tried to see what happens if there is no dropdown and you
overflow...Hmmm...

Also, looks like settings is getting thrown in the bar with dropdown
disabled now, a second one.  Will fix that tomorrow as well.  Either
way, tab colors!