Add All-tab color selector, animations, more code cleanup
Add a selector to settings for when 'per-tab-colors' is disabled.  Add
relevant animations, and a check for never having selected a color
before, at which time it will use the default theme color.

TODO: Reset stored color setting if theme changed, so a theme change
defauts all tab color to theme default.

Fix issue with SHA retrieval if an empty value was stored in settings.

More PHP code cleanup.