Fix "closing modals with escape key", Fix Splash doesn't do anything when mobile
The original method used to close modals on Escape was inadvertently
dismissing the update notification window.  This fixes it and only
allows our pre-defined modals to close.  Pity the BootStrap method for
this doesn't actually work.

Fix splash screen not properly finding and clicking on the tabs when in
mobile mode.