Style fixes, fix "Can't use inputs while mobile" in settings menu...
Fix select alignment for mobile UI.
Fix issue caused by touch-punch overriding inputs on mobile, causing
them to be unusable.  This is a custom fix, and is not present in the
stock touch-punch.
Fix alignment for color labels to match other labels.
Add JS alert function for PHP methods that require it.