Update login.php
First of all fixed the failed login write_log. For me at least it would always write "0" to the log, no matter what.

So it is now working and printing the IP of the failed login. This is much more useful in my opinion, as it can then be used in conjunction with fail2ban, for example.